pyproject.toml
[tool.poetry]
name = "bee-stream-server"
version = "0.4.0"
description = "Bee count analysis with dynamic and static background support"
authors = ["Austin"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.9.0"
websockets = "^12.0"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.4.24"
scipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
